.github/workflows/mingw.yml: preserve artifacts
the amount of verbiage this takes is mind-boggling, and there's
no easy way to cut it down.  the v1 of upload-artifact can only do
one artifact. the v2 is in preview and while it's slated to allow
for multiple artifacts, specified with globs even, but then puts
all the files in a single zip archive.